backend/tg_notify/pyproject.toml
[project]
name = "iubip2025_backend"
version = "0.1.0"
description = "Флагман case"
requires-python = ">=3.11"
dependencies = [
    "redis>=5.2.0,<6.0.0",
    "python-dotenv>=1.1.0",
    "nats-py==2.10.0",
    "aiogram==3.19.0"
]

[project.optional-dependencies]
dev = [
    "setuptools>=75.2.0,<76.0.0",
    "autoflake>=2.3.1,<3.0.0",
    "ruff>=0.6.0,<0.7.0",
    "black>=24.8.0,<25.0.0",
    "pre-commit>=3.8.0",
]
test = [
    "docker>=7.1.0",
    "pytest>=8.3.2",
    "pytest-asyncio>=0.24.0",
    "psycopg2-binary>=2.9.8"
]

[build-system]
requires = ["uv>=0.1.0", "setuptools>=75.2.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 89
target-version = "py311"

exclude = [
    "migrations/",
    "**/__pycache__/",
    "**/.egg-info/",
    "**/alembic/versions/"
]

[tool.ruff.lint]

select = ["ALL"]

ignore = [
    "ARG", "ANN", "D", "EM101", "EM102", "PT001", "PT023",
    "SIM108", "SIM114", "TRY003", "PLW2901", "RET505",
    "PLR0913", "UP038", "TCH001", "SLF001", "COM812",
    "ISC001", "S101", "RUF002", "B008", "E721", "INP001",
    "ERA001", "RUF001", "RUF003", "PTH123", "PTH120", "PTH118",
    "PTH100"
]

[tool.ruff.lint.isort]
no-lines-before = ["local-folder"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = "ignore::DeprecationWarning"
